Copy of relay cap from bitcoin server. enabled an "Authority" setting

+++ src/lib/serviceRelayCap.js
@@ -1,6 +1,8 @@
-const debug = require('./debug');
+var debug = require('./debug');
 const queue = require('./queue');
 const queueMiddleware = require('../middleware/queue');
+
+debug = () => {};
 
 function isEqual(newObject, oldObject) {
   var keys = Object.assign([], Object.keys(newObject), Object.keys(oldObject));
@@ -40,8 +42,11 @@ function isEqual(newObject, oldObject) {
 }
 
 
-module.exports = function (app, localService, remoteService, updateCallback) {
+module.exports = function (app, localService, remoteService, updateCallback, authority) {
 
+  if (typeof updateCallback !== 'function' && authority === void 0 ){
+    authority = updateCallback;
+  }
   var serviceName;
 
   if (typeof remoteService === 'function') {
@@ -77,15 +82,6 @@ module.exports = function (app, localService, remoteService, updateCallback) {
     updateCallback(result);
   });
 
-  remoteService.on('created', (result) => {
-    localService.create(result).then(result => {
-      debug(`Created ${serviceName} on Local`, result);
-    }).catch(error => {
-      debug(`unable to create ${serviceName} on Local`, error);
-    });
-
-  });
-
   localService.on('updated', (result) => {
     queueMiddleware.updateData(serviceName, result, result._id);
     // queue.set(`updated:${serviceName}:${result._id}`, (resolve, reject) => {
@@ -102,36 +98,6 @@ module.exports = function (app, localService, remoteService, updateCallback) {
     updateCallback(result);
   });
 
-  remoteService.on('updated', (result) => {
-    var runCount = 0;
-
-    function runGetLocal() {
-      localService.get(result._id).then(lresult => {
-        if (!isEqual(result, lresult)) {
-          // debug('equal!!', result, '\n===\n', lresult);
-          debug(`About to update ${serviceName}`);
-          localService.update(result._id, {
-            $set: result
-          }).then((result) => {
-            debug(`Updated ${serviceName} on Local`, result);
-          }).catch(error => {
-
-            debug(`Unable to update ${serviceName} on Local`, error);
-          });
-        } else {
-          // debug('equal??', result, '\n===\n', lresult);
-
-        }
-      }).catch(error => {
-        if (runCount < 5) {
-          runCount++;
-          setTimeout(runGetLocal, 1);
-        }
-      });
-    }
-    runGetLocal();
-  });
-
   localService.on('removed', (result) => {
     queueMiddleware.removeData(serviceName, result._id);
     // queue.set(`removed:${serviceName}:${result._id}`, (resolve, reject) => {
@@ -143,11 +109,52 @@ module.exports = function (app, localService, remoteService, updateCallback) {
     // });
   });
 
-  remoteService.on('removed', (result) => {
-    localService.remove(result._id).then((result) => {
-      debug(`Removed ${serviceName} from Local`);
-    }).catch(error => {
-      debug(`Unable to remove ${serviceName} from Local`);
+  if (remoteService && !authority) {
+    remoteService.on('created', (result) => {
+      localService.create(result).then(result => {
+        debug(`Created ${serviceName} on Local`, result);
+      }).catch(error => {
+        debug(`unable to create ${serviceName} on Local`, error);
+      });
+
     });
-  });
+
+    remoteService.on('updated', (result) => {
+      var runCount = 0;
+
+      function runGetLocal() {
+        localService.get(result._id).then(lresult => {
+          if (!isEqual(result, lresult)) {
+            // debug('equal!!', result, '\n===\n', lresult);
+            debug(`About to update ${serviceName}`);
+            localService.update(result._id, {
+              $set: result
+            }).then((result) => {
+              debug(`Updated ${serviceName} on Local`, result);
+            }).catch(error => {
+
+              debug(`Unable to update ${serviceName} on Local`, error);
+            });
+          } else {
+            // debug('equal??', result, '\n===\n', lresult);
+
+          }
+        }).catch(error => {
+          if (runCount < 5) {
+            runCount++;
+            setTimeout(runGetLocal, 1);
+          }
+        });
+      }
+      runGetLocal();
+    });
+
+    remoteService.on('removed', (result) => {
+      localService.remove(result._id).then((result) => {
+        debug(`Removed ${serviceName} from Local`);
+      }).catch(error => {
+        debug(`Unable to remove ${serviceName} from Local`);
+      });
+    });
+  }
 };
